Improved logging system & added channel logs

+++ discord/user/DiscordUserTickets.php
@@ -154,6 +154,7 @@ class DiscordUserTickets
             $channel = $this->plan->discord->getChannel($query->post_channel_id);
 
             if ($channel !== null
+                && $channel->allowText()
                 && $channel->guild_id == $query->post_server_id) {
                 $channel->sendMessage($message);
             }
@@ -374,6 +375,7 @@ class DiscordUserTickets
                         $channel = $this->plan->discord->getChannel($query->created_channel_id);
 
                         if ($channel !== null
+                            && $channel->allowText()
                             && $channel->guild_id == $query->created_channel_server_id) {
                             $this->ignoreDeletion++;
                             $channel->guild->channels->delete(
@@ -741,6 +743,7 @@ class DiscordUserTickets
                     $channel = $this->plan->discord->getChannel($row->created_channel_id);
 
                     if ($channel !== null
+                        && $channel->allowText()
                         && $channel->guild_id == $row->created_channel_server_id) {
                         $this->ignoreDeletion++;
                         $channel->guild->channels->delete($channel);
